fix(storefront): always show current version in version switch | hj / sas
