+++ packages/storefront/src/components/layout/Canvas.tsx
@@ -22,14 +22,11 @@ import { ThemeSelect } from '@/components/common/ThemeSelect';
 import { Search } from '@/components/search/Search';
 import { useDirection } from '@/hooks/useDirection';
 import { useStorefrontColorScheme } from '@/hooks/useStorefrontColorScheme';
+import { useStorefrontVersion } from '@/hooks/useStorefrontVersion';
 import { useTextZoom } from '@/hooks/useTextZoom';
 import type { StorefrontColorScheme } from '@/models/colorScheme';
 import type { StorefrontDirection } from '@/models/dir';
-import { LEGACY_PDS_VERSIONS, type PDSVersionGroup, type Semver } from '@/models/pdsVersion';
 import type { StorefrontTextZoom } from '@/models/textZoom';
-import { fetchPdsVersions } from '@/utils/fetchPdsVersions';
-import { isDevEnvironment } from '@/utils/isDev';
-import { localPorscheDesignSystemVersion } from '@/utils/porscheDesignSystemVersion';
 
 declare global {
   interface Window {
@@ -42,17 +39,7 @@ if (global?.window) {
 }
 
 export const Canvas = ({ children }: PropsWithChildren) => {
-  const [stablePdsReleases, setStablePdsReleases] = useState<string[]>([]);
-
-  useEffect(() => {
-    async function load() {
-      const list = await fetchPdsVersions();
-      setStablePdsReleases(list);
-    }
-
-    load();
-  }, []);
-
+  const { pdsVersion, isOutdatedVersionBannerOpen, setIsIsOutdatedVersionBannerOpen } = useStorefrontVersion();
   const { storefrontColorScheme, setStorefrontColorScheme } = useStorefrontColorScheme();
   const { storefrontDirection, setStorefrontDirection } = useDirection();
   const { storefrontTextZoom, setStorefrontTextZoom } = useTextZoom();
@@ -61,23 +48,7 @@ export const Canvas = ({ children }: PropsWithChildren) => {
   const [isSidebarEndOpen, setIsSidebarEndOpen] = useState(false);
   const [isSearchModalOpen, setIsSearchModalOpen] = useState<boolean>(false);
   const searchInputRef = useRef<HTMLInputElement | null>(null);
-  const [isBannerOpen, setIsBannerOpen] = useState(false);
   const isDesktop = typeof window !== 'undefined' && window.matchMedia(`(min-width: ${breakpointMd}px)`).matches;
-
-  const latestPdsVersion = stablePdsReleases[0] as Semver;
-
-  useEffect(() => {
-    if (!latestPdsVersion) return;
-    if (!isDevEnvironment && localPorscheDesignSystemVersion !== latestPdsVersion) {
-      setIsBannerOpen(true);
-    }
-  }, [latestPdsVersion]);
-
-  const pdsVersion: PDSVersionGroup = {
-    all: [...stablePdsReleases, ...LEGACY_PDS_VERSIONS],
-    current: localPorscheDesignSystemVersion as Semver,
-    latest: latestPdsVersion,
-  };
 
   const onSidebarStartUpdate = (e: CustomEvent<CanvasSidebarStartUpdateEventDetail>) => {
     setIsSidebarStartOpen(e.detail.open);
@@ -127,7 +98,7 @@ export const Canvas = ({ children }: PropsWithChildren) => {
         Porsche Design System
       </Link>
 
-      <PBanner open={isBannerOpen} onDismiss={() => setIsBannerOpen(false)}>
+      <PBanner open={isOutdatedVersionBannerOpen} onDismiss={() => setIsIsOutdatedVersionBannerOpen(false)}>
         <div slot="description" className="flex flex-col gap-fluid-xs">
           You are currently viewing an earlier release of the Porsche Design System.
           <Link href={`https://designsystem.porsche.com/`}>
